+++ FreedomNarratives/utilities/sections/person_event-list.php
@@ -4,7 +4,7 @@
   <table class="table table-bordered" id="dataTable" width="100%" cellspacing="0">
     <thead>
       <tr>
-          <th>Re-Order</th>
+      <?php if($results['security']>=0){ ?><th>Re-Order</th><?php } ?>
           <th>Event ID</th>
           <th>Type</th>
           <th>Start Date</th>
@@ -12,7 +12,7 @@
 
 
           <th> </th>
-          <th> </th>
+          <?php if($results['security']>=0){ ?><th> </th><?php } ?> 
 
       </tr>
     </thead>
@@ -43,15 +43,15 @@ while ($datas = $query->fetch(PDO::FETCH_ASSOC)) {
   $event = $query1->fetch(PDO::FETCH_ASSOC);
 
 echo "<tr>";
-
-if($datas['listorder']==$MIN_listorder_data){
-
-  echo "<td width=\"100\" ! important><h4><a href=\"utilities/commands_external/move_down_event.php?listID=".$datas['listorder']."&personID=".$_GET['personID']."&ID=".$datas['ID']."\" class=\"primary waves-effect waves-light\" ><i class=\"fas fa-arrow-circle-down\"></i></a></h4></td>";
-}elseif ($datas['listorder']==$MAX_listorder_data) {
-  echo "<td width=\"100\" ! important><h4><a href=\"utilities/commands_external/move_up_event.php?listID=".$datas['listorder']."&personID=".$_GET['personID']."&ID=".$datas['ID']."\" class=\"primary waves-effect waves-light\" ><i class=\"fas fa-arrow-circle-up\"></i></a></h4></td>";
-}else{
-  echo "<td width=\"100\" ! important><h4><a href=\"utilities/commands_external/move_up_event.php?listID=".$datas['listorder']."&personID=".$_GET['personID']."&ID=".$datas['ID']."\" class=\"primary waves-effect waves-light\" ><i class=\"fas fa-arrow-circle-up\"></i></a><a href=\"utilities/commands_external/move_down_event.php?listID=".$datas['listorder']."&personID=".$_GET['personID']."&ID=".$datas['ID']."\" class=\"primary waves-effect waves-light\" ><i class=\"fas fa-arrow-circle-down\"></i></a></h4></td>";
-}
+if($results['security']>=0){
+  if($datas['listorder']==$MIN_listorder_data){
+    echo "<td width=\"100\" ! important><h4><a href=\"utilities/commands_external/move_down_event.php?listID=".$datas['listorder']."&personID=".$_GET['personID']."&ID=".$datas['ID']."\" class=\"primary waves-effect waves-light\" ><i class=\"fas fa-arrow-circle-down\"></i></a></h4></td>";
+  }elseif ($datas['listorder']==$MAX_listorder_data) {
+    echo "<td width=\"100\" ! important><h4><a href=\"utilities/commands_external/move_up_event.php?listID=".$datas['listorder']."&personID=".$_GET['personID']."&ID=".$datas['ID']."\" class=\"primary waves-effect waves-light\" ><i class=\"fas fa-arrow-circle-up\"></i></a></h4></td>";
+  }else{
+    echo "<td width=\"100\" ! important><h4><a href=\"utilities/commands_external/move_up_event.php?listID=".$datas['listorder']."&personID=".$_GET['personID']."&ID=".$datas['ID']."\" class=\"primary waves-effect waves-light\" ><i class=\"fas fa-arrow-circle-up\"></i></a><a href=\"utilities/commands_external/move_down_event.php?listID=".$datas['listorder']."&personID=".$_GET['personID']."&ID=".$datas['ID']."\" class=\"primary waves-effect waves-light\" ><i class=\"fas fa-arrow-circle-down\"></i></a></h4></td>";
+  }
+} 
 echo "<td width=\"130\" ! important>" . $event['UI'] . "</td>";
 if($event['Event_Type']!='0'){
   if(fnmatch("*Connection Only*",$datas['Type'])){
@@ -72,7 +72,9 @@ echo "<td width=\"154\" ! important>" . $event['Field1'] . "</td>";
 
 
 echo "<td width=\"154\" ! important><a href=\"event_edit.php?personID=".$_GET['personID']."&eventID=".$event['eventID']."\" class=\"btn btn-success btn-block waves-effect waves-light\">Update</a></td>";
-echo "<td width=\"154\" ! important><a href=\"utilities/commands_external/delete_event.php?eventID=".$event['eventID']."\" class=\"btn btn-secondary btn-block waves-effect waves-light\" title=\"Delete Event From the Database. To remove only for this individual click 'Update'.\">Delete</a></td>";
+if($results['security']>=0){
+  echo "<td width=\"154\" ! important><a href=\"utilities/commands_external/delete_event.php?eventID=".$event['eventID']."\" class=\"btn btn-secondary btn-block waves-effect waves-light\" title=\"Delete Event From the Database. To remove only for this individual click 'Update'.\">Delete</a></td>";
+}
